Relocate apps.resource
