+++ bcs-app/backend/uniapps/resource/__init__.py
@@ -1,0 +1,1 @@
+"""管理统一应用资源，包括 ConfigMap、Secret 等"""
