+++ bcs-app/backend/uniapps/resource/urls.py
