+++ bcs-app/backend/uniapps/resource/views.py
