+++ bcs-app/backend/uniapps/resource/views_ingress.py
@@ -23,8 +23,8 @@ from backend.apps.application.utils import APIResponse
 from backend.apps.configuration.k8s.serializers import K8sIngressSLZ
 from backend.apps.constants import ClusterType
 from backend.apps.instance.constants import K8S_INGRESS_SYS_CONFIG
-from backend.apps.resource.views import ResourceOperate
 from backend.components.bcs import k8s
+from backend.uniapps.resource.views import ResourceOperate
 from backend.utils.errcodes import ErrorCode
 from backend.utils.error_codes import error_codes
